requirements.txt
streamlit
openai
python-dotenv
pandas
sqlite-utils
requests
